requirements.txt
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.13
dj-database-url==0.5.0
django-allauth==0.41.0
django-countries==6.1.2
django-crispy-forms==1.9.0
django-extensions==2.2.9
django-heroku==0.3.1
django-seed==0.2.2
factory-boy==2.12.0
Faker==4.0.3
gunicorn==20.0.4
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
mysqlclient==1.4.6
oauthlib==3.1.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
prompt-toolkit==3.0.5
psycopg2==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
whitenoise==5.0.1
